+++ lib/AppInfo/Application.php
@@ -9,6 +9,7 @@ use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
 use OCP\AppFramework\Bootstrap\IRegistrationContext;
+use OCP\IConfig;
 use OCP\ISession;
 use OCP\IUserSession;
 use OCP\Util;
@@ -22,7 +23,13 @@ class Application extends App implements IBootstrap {
 	}
 
 	public function register(IRegistrationContext $context): void {
-		$context->registerAlternativeLogin(HitobitoLogin::class);
+		$config = $this->getContainer()->get(IConfig::class);
+		$generalSettings = (array)$config->getSystemValue(Application::APP_ID);
+
+		// TODO: Create settings validator and use it here
+		if ($generalSettings['base_url'] && $generalSettings['client_id'] && $generalSettings['client_secret'] && $generalSettings['login_button_text']) {
+			$context->registerAlternativeLogin(HitobitoLogin::class);
+		}
 	}
 
 	public function boot(IBootContext $context): void {